# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MeetMind API",
    "version": "1.0.0",
    "description": "REST API for MeetMind, an AI meeting intelligence app. Mobile clients authenticate at POST /api/mobile-auth/login and send the returned token as `Authorization: Bearer <token>`. All endpoints marked with bearerAuth require that token."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/upload/presigned": {
      "post": {
        "operationId": "createPresignedUpload",
        "summary": "Get a presigned URL for uploading a file to cloud storage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fileName",
                  "contentType"
                ],
                "properties": {
                  "fileName": {
                    "type": "string"
                  },
                  "contentType": {
                    "type": "string"
                  },
                  "isPublic": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presigned upload URL and the resulting storage path",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "uploadUrl",
                    "cloud_storage_path"
                  ],
                  "properties": {
                    "uploadUrl": {
                      "type": "string"
                    },
                    "cloud_storage_path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/transcribe": {
      "post": {
        "operationId": "transcribeAudio",
        "summary": "Transcribe base64-encoded audio (streaming plain-text response)",
        "description": "Returns a streamed text/plain body containing OpenAI-style SSE chunks from the transcription model; concatenate the streamed content client-side.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "audioBase64"
                ],
                "properties": {
                  "audioBase64": {
                    "type": "string",
                    "description": "Base64-encoded audio data (no data URI prefix)."
                  },
                  "fileName": {
                    "type": "string",
                    "default": "audio.webm"
                  },
                  "mimeType": {
                    "type": "string",
                    "default": "audio/webm"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Streamed transcription text",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token obtained from POST /api/mobile-auth/login."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  }
}